Move the XBee configurator class to a separate file
This commit renames the `xbee` folder to `zigbee` as the XBee library is
also called `xbee`, causing a name conflict. Additionally, this commit
removes an unused `time` import.
